qcacld-3.0: Fix the incorrect data type for vdev_id
vdev_id data type used in csr_scan_abort_mac_scan()
is declared as uint8_t and is compared with INVALID_VDEV_ID,
a 32 bit integer, which can never be true.

CRS-Fixed: 2315003

+++ core/sme/src/csr/csr_api_scan.c
@@ -1809,7 +1809,7 @@ void csr_set_cfg_scan_control_list(tpAniSirGlobal pMac, uint8_t *countryCode,
 	} /* AllocateMemory */
 }
 
-QDF_STATUS csr_scan_abort_mac_scan(tpAniSirGlobal mac_ctx, uint8_t vdev_id,
+QDF_STATUS csr_scan_abort_mac_scan(tpAniSirGlobal mac_ctx, uint32_t vdev_id,
 				   uint32_t scan_id)
 {
 	struct scan_cancel_request *req;
